Add shared component for NoDocumentationBanner
