+++ coral/src/app/features/components/documentation/components/NoDocumentationBanner.test.tsx
@@ -1,19 +1,20 @@
 import { cleanup, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { NoDocumentationBanner } from "src/app/features/topics/details/documentation/components/NoDocumentationBanner";
+import { NoDocumentationBanner } from "src/app/features/components/documentation/components/NoDocumentationBanner";
 
 const testAddDocumentation = jest.fn();
 
 describe("NoDocumentationBanner", () => {
   const user = userEvent.setup();
 
-  describe("if user is topic owner", () => {
-    describe("shows all necessary elements", () => {
+  describe("if user is entity owner", () => {
+    describe("shows all necessary elements for entity topic", () => {
       beforeAll(() => {
         render(
           <NoDocumentationBanner
+            entity={"topic"}
             addDocumentation={testAddDocumentation}
-            isUserTopicOwner={true}
+            isUserOwner={true}
           />
         );
       });
@@ -40,12 +41,47 @@ describe("NoDocumentationBanner", () => {
       });
     });
 
+    describe("shows all necessary elements for entity connector", () => {
+      beforeAll(() => {
+        render(
+          <NoDocumentationBanner
+            entity={"connector"}
+            addDocumentation={testAddDocumentation}
+            isUserOwner={true}
+          />
+        );
+      });
+
+      afterAll(cleanup);
+
+      it("shows headline No Documentation", () => {
+        const headline = screen.getByRole("heading", {
+          name: "No readme available",
+        });
+        expect(headline).toBeVisible();
+      });
+
+      it("shows information that no readme is available for this connector", () => {
+        const infoText = screen.getByText(
+          "Add a readme to give your team essential information, guidelines, and context about the connector."
+        );
+
+        expect(infoText).toBeVisible();
+      });
+
+      it("shows button to add a readme", () => {
+        const button = screen.getByRole("button", { name: "Add readme" });
+        expect(button).toBeVisible();
+      });
+    });
+
     describe("triggers a addDocumentation event", () => {
       beforeEach(() => {
         render(
           <NoDocumentationBanner
             addDocumentation={testAddDocumentation}
-            isUserTopicOwner={true}
+            entity={"topic"}
+            isUserOwner={true}
           />
         );
       });
@@ -64,13 +100,14 @@ describe("NoDocumentationBanner", () => {
     });
   });
 
-  describe("if user is not topic owner", () => {
+  describe("if user is not entity owner", () => {
     describe("shows all necessary elements", () => {
       beforeAll(() => {
         render(
           <NoDocumentationBanner
+            entity={"topic"}
             addDocumentation={testAddDocumentation}
-            isUserTopicOwner={false}
+            isUserOwner={false}
           />
         );
       });
@@ -86,7 +123,7 @@ describe("NoDocumentationBanner", () => {
 
       it("shows no information that no readme is available for this topic", () => {
         const infoText = screen.queryByText(
-          "Add a readme to give your team essential information, guidelines, and context about the topic."
+          /Add a readme to give your team essential information/
         );
         expect(infoText).not.toBeInTheDocument();
       });
