+++ coral/src/app/features/components/documentation/components/NoDocumentationBanner.tsx
@@ -2,11 +2,25 @@ import { EmptyState, EmptyStateLayout } from "@aivenio/aquarium";
 import illustration from "src/app/images/documentation-illustration.svg";
 
 type NoDocumentationBannerProps = {
+  entity: "connector" | "topic";
   addDocumentation: () => void;
+  isUserOwner: boolean;
 };
 function NoDocumentationBanner({
+  entity,
   addDocumentation,
+  isUserOwner,
 }: NoDocumentationBannerProps) {
+  if (!isUserOwner) {
+    return (
+      <EmptyState
+        title={"No readme available"}
+        image={illustration}
+        layout={EmptyStateLayout.CenterHorizontal}
+      />
+    );
+  }
+
   return (
     <EmptyState
       title={"No readme available"}
@@ -17,8 +31,7 @@ function NoDocumentationBanner({
         onClick: addDocumentation,
       }}
     >
-      Add a readme to give your team essential information, guidelines, and
-      context about the connector.
+      {`Add a readme to give your team essential information, guidelines, and context about the ${entity}.`}
     </EmptyState>
   );
 }
